requirements.txt
fastapi==0.65.2
pymongo==3.11.4
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
starlette==0.14.2
uvicorn==0.14.0
passlib==1.7.4
websockets==9.1
pyjwt==2.1.0
jproperties==2.1.0
docker==5.0.0
requests==2.26.0
redis==3.5.3
google-cloud-dialogflow==2.4.0
fastapi-utils==0.2.1
apscheduler